requirements.txt
schedule
python-json-logger
selenium
stem
python-dotenv